refactor(blog-client): :art: refactored queryLogs
